Windows fix (sadly)

+++ src/main.c
@@ -8,7 +8,7 @@ int main(int argc, char *argv[])
 	init_era(era);
 
 	size_t bytes_read = read_file("test.bin", era);
-	printf("Read %zd bytes\n", bytes_read);
+	printf("Read %lu bytes\n", (unsigned long)bytes_read);
 
 	free_era(era);
 	return 0;
